+++ tilebox-workflows/tilebox/workflows/runner/task_runner.py
@@ -403,7 +403,7 @@ class PollingTaskRunner:
 
 
 class TaskRunner:
-    def __init__(  # noqa: PLR0913
+    def __init__(  # noqa: PLR0913, PLR0917
         self,
         service: TaskService,
         cluster: str,